refactor(dashboard): remove dead code and dedupe abstractions
Resolve verified LLM-artifact findings (#8-13, #25):
- Delete unused useWorkflows hook + its test (no production consumer;
  WorkflowsPage uses useLoaderData directly) (#11, #25)
- brainstorm.ts: drop local createTimeoutSignal, import shared one from
  api/utils (matches client.ts/settings.ts) (#12)
- Consolidate duplicated test files into their __tests__/ siblings,
  porting unique coverage first (no coverage loss):
  - api/client.test.ts createWorkflow cases -> __tests__/client.test.ts (#8)
  - StatusBadge running aria-label/data-status -> __tests__ (#9)
  - ApprovalControls: merge both sets into __tests__ (#10)
- Extract shared useCopyFeedback hook from CopyButton and
  MessageMetadata (#13)
