+++ dashboard/src/components/brainstorm/CopyButton.tsx
@@ -1,7 +1,7 @@
-import { useState, useCallback, useRef, useEffect } from "react";
-import { cn, copyToClipboard } from "@/lib/utils";
+import { cn } from "@/lib/utils";
 import { Copy, Check } from "lucide-react";
 import { Button } from "@/components/ui/button";
+import { useCopyFeedback } from "@/hooks/useCopyFeedback";
 
 interface CopyButtonProps {
   content: string;
@@ -16,29 +16,7 @@ interface CopyButtonProps {
  * - Uses iOS-compatible clipboard handling
  */
 export function CopyButton({ content, className }: CopyButtonProps) {
-  const [copied, setCopied] = useState(false);
-  const timeoutRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-
-  const handleCopy = useCallback(async () => {
-    const success = await copyToClipboard(content);
-    if (success) {
-      setCopied(true);
-      // Clear any existing timeout
-      if (timeoutRef.current) {
-        clearTimeout(timeoutRef.current);
-      }
-      timeoutRef.current = setTimeout(() => setCopied(false), 2000);
-    }
-  }, [content]);
-
-  // Clean up timeout on unmount to prevent memory leak
-  useEffect(() => {
-    return () => {
-      if (timeoutRef.current) {
-        clearTimeout(timeoutRef.current);
-      }
-    };
-  }, []);
+  const { copied, handleCopy } = useCopyFeedback(content);
 
   return (
     <Button
